refactor: title to shared
