+++ src/features/home/about/about.tsx
@@ -1,7 +1,8 @@
-import { Paragraph, Subtitle, Title } from '@/app/components';
+import { Paragraph, Subtitle } from '@/app/components';
 
 import image from '@/assets/about.webp';
 import Image from '@/shared/image';
+import { Title } from '@/shared/title';
 
 import './about.scss';
