+++ src/features/home/contribute/contribute.tsx
@@ -1,7 +1,8 @@
-import { OptionItem, type OptionItemProps, Paragraph, Subtitle, Title } from '@/app/components';
+import { OptionItem, type OptionItemProps, Paragraph, Subtitle } from '@/app/components';
 
 import image from '@/assets/contribute.webp';
 import Image from '@/shared/image';
+import { Title } from '@/shared/title';
 
 import './contribute.scss';
