+++ src/features/home/pictures/pictures.tsx
@@ -1,6 +1,6 @@
 // requires a loader
 import { Carousel } from 'react-responsive-carousel';
-import { Paragraph, SocialMedia, SocialMediaProps, Title } from '@/app/components';
+import { Paragraph, SocialMedia, SocialMediaProps } from '@/app/components';
 
 import photo1 from '@/assets/photo-1.webp';
 import photo2 from '@/assets/photo-2.webp';
@@ -13,6 +13,7 @@ import photo8 from '@/assets/photo-8.webp';
 import photo9 from '@/assets/photo-9.webp';
 import { FacebookIcon, InstagramIcon } from '@/icons';
 import Image from '@/shared/image';
+import { Title } from '@/shared/title';
 
 import 'react-responsive-carousel/lib/styles/carousel.min.css';
 import './pictures.scss';
