+++ src/features/school/study-with-us/study-with-us.tsx
@@ -1,9 +1,10 @@
 import { studyOptions } from './constants';
-import { OptionItem, Paragraph, Subtitle, Title } from '@/app/components';
+import { OptionItem, Paragraph, Subtitle } from '@/app/components';
 
 import image from '@/assets/rs-school.webp';
 import Image from '@/shared/image';
 import { SectionLabel } from '@/shared/section-label';
+import { Title } from '@/shared/title';
 
 import './study-with-us.scss';
